requirements.txt
streamlit
streamlit-authenticator
pyyaml
langchain-google-genai
python-dotenv
pandas
openpyxl
markdown-pdf
plotly
